+++ .storybook/preview.tsx
@@ -3,13 +3,19 @@ import type { Decorator, Preview } from "@storybook/react";
 import React from "react";
 import { theme } from "../src/themes/elcom";
 import "./preview.css";
+import { i18n } from "../src/locales/locales";
+import { I18nProvider } from "@lingui/react";
 
 const withAppProviders: Decorator = (Story: any) => {
   return (
-    <ThemeProvider theme={theme}>
-      <CssBaseline />
-      <Story />
-    </ThemeProvider>
+    <div data-storybook-state="loaded">
+      <I18nProvider i18n={i18n}>
+        <ThemeProvider theme={theme}>
+          <CssBaseline />
+          <Story />
+        </ThemeProvider>
+      </I18nProvider>
+    </div>
   );
 };